requirements.txt
streamlit
pathlib
google-generativeai
python-dotenv